requirements.txt
requests>=2.32.0
google-generativeai>=0.3.0
transformers>=4.51.3
torch>=2.0.0
accelerate>=0.27.0
python-dotenv~=1.1.0